feat: 컴포넌트 분리 및 submit 설정
